Split InstrumentTrackView and InstrumentTrackWindow off of InstrumentTrack
