+++ include/InstrumentTrack.h
@@ -1,6 +1,6 @@
 /*
- * InstrumentTrack.h - declaration of class InstrumentTrack, a track + window
- *                     which holds an instrument-plugin
+ * InstrumentTrack.h - declaration of class InstrumentTrack, a track which
+ *                     holds an instrument-plugin
  *
  * Copyright (c) 2004-2014 Tobias Doerffel <tobydox/at/users.sourceforge.net>
  *
@@ -326,197 +326,5 @@ private:
 	friend class MidiCCRackView;
 
 } ;
-
-
-
-
-class InstrumentTrackView : public TrackView
-{
-	Q_OBJECT
-public:
-	InstrumentTrackView( InstrumentTrack * _it, TrackContainerView* tc );
-	virtual ~InstrumentTrackView();
-
-	InstrumentTrackWindow * getInstrumentTrackWindow();
-
-	InstrumentTrack * model()
-	{
-		return castModel<InstrumentTrack>();
-	}
-
-	const InstrumentTrack * model() const
-	{
-		return castModel<InstrumentTrack>();
-	}
-
-	static InstrumentTrackWindow * topLevelInstrumentTrackWindow();
-
-	QMenu * midiMenu()
-	{
-		return m_midiMenu;
-	}
-
-	// Create a menu for assigning/creating channels for this track
-	QMenu * createFxMenu( QString title, QString newFxLabel ) override;
-
-
-protected:
-	void dragEnterEvent( QDragEnterEvent * _dee ) override;
-	void dropEvent( QDropEvent * _de ) override;
-
-
-private slots:
-	void toggleInstrumentWindow( bool _on );
-	void toggleMidiCCRack();
-	void activityIndicatorPressed();
-	void activityIndicatorReleased();
-
-	void midiInSelected();
-	void midiOutSelected();
-	void midiConfigChanged();
-
-	void assignFxLine( int channelIndex );
-	void createFxLine();
-
-	void handleConfigChange(QString cls, QString attr, QString value);
-
-
-private:
-	InstrumentTrackWindow * m_window;
-
-	// widgets in track-settings-widget
-	TrackLabelButton * m_tlb;
-	Knob * m_volumeKnob;
-	Knob * m_panningKnob;
-	FadeButton * m_activityIndicator;
-
-	QMenu * m_midiMenu;
-
-	QAction * m_midiInputAction;
-	QAction * m_midiOutputAction;
-
-	std::unique_ptr<MidiCCRackView> m_midiCCRackView;
-
-	QPoint m_lastPos;
-
-	FadeButton * getActivityIndicator() override
-	{
-		return m_activityIndicator;
-	}
-
-	friend class InstrumentTrackWindow;
-
-} ;
-
-
-
-
-class InstrumentTrackWindow : public QWidget, public ModelView,
-								public SerializingObjectHook
-{
-	Q_OBJECT
-public:
-	InstrumentTrackWindow( InstrumentTrackView * _tv );
-	virtual ~InstrumentTrackWindow();
-
-	// parent for all internal tab-widgets
-	TabWidget * tabWidgetParent()
-	{
-		return m_tabWidget;
-	}
-
-	InstrumentTrack * model()
-	{
-		return castModel<InstrumentTrack>();
-	}
-
-	const InstrumentTrack * model() const
-	{
-		return castModel<InstrumentTrack>();
-	}
-
-	void setInstrumentTrackView( InstrumentTrackView * _tv );
-
-	InstrumentTrackView *instrumentTrackView()
-	{
-		return m_itv;
-	}
-
-
-	PianoView * pianoView()
-	{
-		return m_pianoView;
-	}
-
-	static void dragEnterEventGeneric( QDragEnterEvent * _dee );
-
-	void dragEnterEvent( QDragEnterEvent * _dee ) override;
-	void dropEvent( QDropEvent * _de ) override;
-
-
-public slots:
-	void textChanged( const QString & _new_name );
-	void toggleVisibility( bool _on );
-	void updateName();
-	void updateInstrumentView();
-
-
-protected:
-	// capture close-events for toggling instrument-track-button
-	void closeEvent( QCloseEvent * _ce ) override;
-	void focusInEvent( QFocusEvent * _fe ) override;
-
-	void saveSettings( QDomDocument & _doc, QDomElement & _this ) override;
-	void loadSettings( const QDomElement & _this ) override;
-
-
-protected slots:
-	void saveSettingsBtnClicked();
-	void viewNextInstrument();
-	void viewPrevInstrument();
-
-private:
-	void modelChanged() override;
-	void viewInstrumentInDirection(int d);
-	//! adjust size of any child widget of the main tab
-	//! required to keep the old look when using a variable sized tab widget
-	void adjustTabSize(QWidget *w);
-
-	InstrumentTrack * m_track;
-	InstrumentTrackView * m_itv;
-
-	// widgets on the top of an instrument-track-window
-	QLineEdit * m_nameLineEdit;
-	LeftRightNav * m_leftRightNav;
-	Knob * m_volumeKnob;
-	Knob * m_panningKnob;
-	Knob * m_pitchKnob;
-	QLabel * m_pitchLabel;
-	LcdSpinBox* m_pitchRangeSpinBox;
-	QLabel * m_pitchRangeLabel;
-	FxLineLcdSpinBox * m_effectChannelNumber;
-
-
-
-	// tab-widget with all children
-	TabWidget * m_tabWidget;
-	PluginView * m_instrumentView;
-	InstrumentSoundShapingView * m_ssView;
-	InstrumentFunctionNoteStackingView* m_noteStackingView;
-	InstrumentFunctionArpeggioView* m_arpeggioView;
-	InstrumentMidiIOView * m_midiView;
-	EffectRackView * m_effectView;
-	InstrumentMiscView *m_miscView;
-
-
-	// test-piano at the bottom of every instrument-settings-window
-	PianoView * m_pianoView;
-
-	friend class InstrumentView;
-	friend class InstrumentTrackView;
-
-} ;
-
-
 
 #endif
