+++ src/gui/CMakeLists.txt
@@ -19,6 +19,8 @@ SET(LMMS_SRCS
 	gui/FxMixerView.cpp
 	gui/GuiApplication.cpp
 	gui/InstrumentView.cpp
+	gui/InstrumentTrackView.cpp
+	gui/InstrumentTrackWindow.cpp
 	gui/LfoControllerDialog.cpp
 	gui/LmmsPalette.cpp
 	gui/LmmsStyle.cpp
