+++ src/gui/InstrumentView.cpp
@@ -27,6 +27,7 @@
 #include "InstrumentView.h"
 #include "embed.h"
 #include "InstrumentTrack.h"
+#include "InstrumentTrackWindow.h"
 #include "StringPairDrag.h"
