+++ src/gui/MainWindow.cpp
@@ -51,7 +51,8 @@
 #include "FxMixerView.h"
 #include "GuiApplication.h"
 #include "ImportFilter.h"
-#include "InstrumentTrack.h"
+#include "InstrumentTrackView.h"
+#include "InstrumentTrackWindow.h"
 #include "MicrotunerConfig.h"
 #include "PianoRoll.h"
 #include "PluginBrowser.h"
